feat: generated utxo (#28)
* feat: completed function for generating utxo

* refactor: used block_persistence for the persistence layer

* fix: fixed the block and tx sync

* fix: fixed test errors

* fix: fixed block tests

* fix: koko

+++ src/model/block/block_persistence.c
@@ -1,5 +1,6 @@
 #include "block_persistence.h"
 
+#include <glib.h>
 #include <string.h>
 
 #include "model/transaction/transaction_persistence.h"
@@ -8,6 +9,33 @@
 #include "utils/mysql_util.h"
 
 #define LOG_SCOPE "block_persistence"
+
+static GHashTable *g_global_block_table;  // The global block table that maps block header hash to the block.
+
+/**
+ * Free the memory space of a block.
+ * @param block_destroy The block to be destroyed.
+ * @author Junjian Chen
+ */
+void destroy_block(block *block_destroy) {
+    if (PERSISTENCE_MODE == PERSISTENCE_RAM) {
+        free(block_destroy->header);
+        free(block_destroy);
+    } else if (PERSISTENCE_MODE == PERSISTENCE_MYSQL) {
+        // TODO: Delete an entry of block here.
+    }
+}
+
+/*
+ * -----------------------------------------------------------
+ * Helper Methods
+ * -----------------------------------------------------------
+ */
+
+void free_g_global_block_table_entry(void *block_id, void *blk, void *user_data) {
+    free(block_id);
+    destroy_block(blk);
+}
 
 /*
  * -----------------------------------------------------------
@@ -46,6 +74,8 @@ bool initialize_block_persistence() {
             ");";
         return mysql_create_table(sql_query);
     } else if (PERSISTENCE_MODE == PERSISTENCE_RAM) {
+        g_global_block_table = g_hash_table_new(g_str_hash, g_str_equal);
+        return true;
     }
 
     return false;
@@ -58,58 +88,86 @@ bool initialize_block_persistence() {
  * @auhtor Luke E
  */
 bool save_block(block *bl) {
-    int temp_sql_query_size = 10000;
+    if (PERSISTENCE_MODE == PERSISTENCE_MYSQL) {
+        int temp_sql_query_size = 10000;
 
-    // Save the block header in table blockheader.
-    char sql_query[temp_sql_query_size];
-    memset(sql_query, 0, temp_sql_query_size);
-    sprintf(sql_query,
-            "set @txn_count := %d;\n"
-            "insert into block(block_id, txn_count)\n"
-            "values (NULL, @txn_count);\n"
-            "set @block_h_id = LAST_INSERT_ID();\n",
-            bl->txn_count);
-    if (!mysql_insert(sql_query)) {
-        general_log(LOG_SCOPE, LOG_ERROR, "Failed to insert block.");
-        return false;
-    };
-    unsigned long block_id = mysql_get_last_updated_id();
-    memset(sql_query, '\0', temp_sql_query_size);
+        // Save the block header in table blockheader.
+        char sql_query[temp_sql_query_size];
+        memset(sql_query, 0, temp_sql_query_size);
+        sprintf(sql_query,
+                "set @txn_count := %d;\n"
+                "insert into block(block_id, txn_count)\n"
+                "values (NULL, @txn_count);\n"
+                "set @block_h_id = LAST_INSERT_ID();\n",
+                bl->txn_count);
+        if (!mysql_insert(sql_query)) {
+            general_log(LOG_SCOPE, LOG_ERROR, "Failed to insert block.");
+            return false;
+        };
+        memset(sql_query, '\0', temp_sql_query_size);
 
-    // Insert block header.
-    block_header *current_header = bl->header;
-    sprintf(sql_query,
-            "set @version = %d;\n"
-            "set @prev_block_header_hash = '%s';\n"
-            "set @merkle_root_hash = '%s';\n"
-            "set @block_header_hash = '%s';\n"
-            "set @time = %u;\n"
-            "set @nBits = %u;\n"
-            "set @nonce= %u;\n"
-            "insert into block_header(block_h_id, version, block_header_hash, prev_block_header_hash, merkle_root_hash, time, nBits, nonce)\n"
-            "values (@block_h_id, @version, @block_header_hash, @prev_block_header_hash, @merkle_root_hash, @time, @nBits, @nonce);",
-            current_header->version,
-            current_header->prev_block_header_hash,
-            current_header->merkle_root_hash,
-            hash_block_header(current_header),
-            current_header->time,
-            current_header->nBits,
-            current_header->nonce);
-    if (!mysql_insert(sql_query)) {
-        general_log(LOG_SCOPE, LOG_ERROR, "Failed to insert block header.");
-        return false;
+        // Insert block header.
+        block_header *current_header = bl->header;
+        sprintf(sql_query,
+                "set @version = %d;\n"
+                "set @prev_block_header_hash = '%s';\n"
+                "set @merkle_root_hash = '%s';\n"
+                "set @block_header_hash = '%s';\n"
+                "set @time = %u;\n"
+                "set @nBits = %u;\n"
+                "set @nonce= %u;\n"
+                "insert into block_header(block_h_id, version, block_header_hash, prev_block_header_hash, merkle_root_hash, time, nBits, nonce)\n"
+                "values (@block_h_id, @version, @block_header_hash, @prev_block_header_hash, @merkle_root_hash, @time, @nBits, @nonce);",
+                current_header->version,
+                current_header->prev_block_header_hash,
+                current_header->merkle_root_hash,
+                hash_block_header(current_header),
+                current_header->time,
+                current_header->nBits,
+                current_header->nonce);
+        if (!mysql_insert(sql_query)) {
+            general_log(LOG_SCOPE, LOG_ERROR, "Failed to insert block header.");
+            return false;
+        }
+        memset(sql_query, '\0', temp_sql_query_size);
+
+        // Update the block ID for all associated transactions.
+        unsigned long block_id = get_block_id_in_database(bl);
+        for (int i = 0; i < bl->txn_count; i++) {
+            transaction *current_transaction = bl->txns[i];
+            char *txid = get_transaction_txid(current_transaction);
+            update_transaction_block_id(block_id, txid);
+            free(txid);
+        }
+
+        return true;
+    } else if (PERSISTENCE_MODE == PERSISTENCE_RAM) {
+        char *cur_block_hash = hash_block_header(bl->header);
+        g_hash_table_insert(g_global_block_table, cur_block_hash, bl);
+        return true;
     }
-    memset(sql_query, '\0', temp_sql_query_size);
 
-    // Update the block ID for all associated transactions.
-    for (int i = 0; i < bl->txn_count; i++) {
-        transaction *current_transaction = bl->txns[i];
-        char *txid = get_transaction_txid(current_transaction);
-        update_transaction_block_id(block_id, txid);
-        free(txid);
+    return false;
+}
+
+/**
+ * Get the block ID of a specific block in the MySQL database.
+ * @param block The specific block.
+ * @return Its block ID.
+ */
+unsigned long get_block_id_in_database(block *block) {
+    char *header_hash = hash_block_header(block->header);
+    char sql_query[1000];
+    sprintf(sql_query, "select block_h_id from block_header where block_header_hash='%s';", header_hash);
+    MYSQL_RES *res = mysql_read(sql_query);
+
+    MYSQL_ROW row;
+    unsigned long block_header_id;
+    while ((row = mysql_fetch_row(res))) {
+        block_header_id = atoi(row[0]);
     }
-
-    return true;
+    free(header_hash);
+    return block_header_id;
 }
 
 /**
@@ -118,7 +176,7 @@ bool save_block(block *bl) {
  * @return True for exists and false otherwise.
  * @author Luke E
  */
-bool does_blocK_exist(char *block_header_hash) {
+bool does_block_exist(char *block_header_hash) {
     char sql_query[1000];
     memset(sql_query, '\0', 1000);
     sprintf(sql_query, "select * from block_header where block_header_hash='%s';", block_header_hash);
@@ -135,63 +193,69 @@ bool does_blocK_exist(char *block_header_hash) {
  * @author Luke E
  */
 block *get_block(char *block_header_hash) {
-    block *b = (block *)malloc(sizeof(block));
-    memset(b, 0, sizeof(block));
-    b->header = (block_header *)(malloc(sizeof(block_header)));
-    memset(b->header, 0, sizeof(block_header));
-    int temp_sql_query_size = 10000;
-    char sql_query[temp_sql_query_size];
+    if (PERSISTENCE_MODE == PERSISTENCE_MYSQL) {
+        block *b = (block *)malloc(sizeof(block));
+        memset(b, 0, sizeof(block));
+        b->header = (block_header *)(malloc(sizeof(block_header)));
+        memset(b->header, 0, sizeof(block_header));
+        int temp_sql_query_size = 10000;
+        char sql_query[temp_sql_query_size];
 
-    // Read block header.
-    sprintf(sql_query, "select * from block_header where block_header_hash='%s';", block_header_hash);
-    MYSQL_RES *res = mysql_read(sql_query);
-    MYSQL_ROW row = mysql_fetch_row(res);
-    unsigned long block_id;
-    block_id = atoi(row[0]);
-    b->header->version = atoi(row[1]);
-    memcpy(b->header->prev_block_header_hash, row[2], 64);
-    memcpy(b->header->merkle_root_hash, row[3], 64);
-    b->header->time = atoi(row[4]);
-    b->header->nBits = atoi(row[5]);
-    b->header->nonce = atoi(row[6]);
-    mysql_free_result(res);
-    memset(sql_query, 0, temp_sql_query_size);
+        // Read block header.
+        sprintf(sql_query, "select * from block_header where block_header_hash='%s';", block_header_hash);
+        MYSQL_RES *res = mysql_read(sql_query);
+        MYSQL_ROW row = mysql_fetch_row(res);
+        unsigned long block_id;
+        block_id = atoi(row[0]);
+        b->header->version = atoi(row[1]);
+        memcpy(b->header->prev_block_header_hash, row[2], 64);
+        memcpy(b->header->merkle_root_hash, row[3], 64);
+        b->header->time = atoi(row[4]);
+        b->header->nBits = atoi(row[5]);
+        b->header->nonce = atoi(row[6]);
+        mysql_free_result(res);
+        memset(sql_query, 0, temp_sql_query_size);
 
-    // Read block.
-    sprintf(sql_query, "select * from block where block_id=%lu;", block_id);
-    res = mysql_read(sql_query);
-    row = mysql_fetch_row(res);
-    b->txn_count = atoi(row[1]);
-    mysql_free_result(res);
-    memset(sql_query, 0, temp_sql_query_size);
+        // Read block.
+        sprintf(sql_query, "select * from block where block_id=%lu;", block_id);
+        res = mysql_read(sql_query);
+        row = mysql_fetch_row(res);
+        b->txn_count = atoi(row[1]);
+        mysql_free_result(res);
+        memset(sql_query, 0, temp_sql_query_size);
 
-    // Read associated transactions.
-    char **txids = (char **)malloc(b->txn_count * sizeof(char *));
-    memset(txids, 0, b->txn_count);
-    for (int i = 0; i < b->txn_count; i++) {
-        txids[i] = (char *)malloc(65);
-        memset(txids[i], 0, 65);
+        // Read associated transactions.
+        char **txids = (char **)malloc(b->txn_count * sizeof(char *));
+        memset(txids, 0, b->txn_count);
+        for (int i = 0; i < b->txn_count; i++) {
+            txids[i] = (char *)malloc(65);
+            memset(txids[i], 0, 65);
+        }
+
+        sprintf(sql_query, "select txid from transaction where block_id=%lu;", block_id);
+        res = mysql_read(sql_query);
+        int tx_count = 0;
+        while ((row = mysql_fetch_row(res))) {
+            memcpy(txids[tx_count], row[0], 64);
+            tx_count++;
+        }
+        b->txns = (transaction **)malloc(b->txn_count * sizeof(transaction *));
+        for (int i = 0; i < b->txn_count; i++) {
+            b->txns[i] = get_transaction(txids[i]);
+        }
+
+        mysql_free_result(res);
+        memset(sql_query, 0, temp_sql_query_size);
+        for (int i = 0; i < b->txn_count; i++) free(txids[i]);
+        free(txids);
+
+        return b;
+    } else if (PERSISTENCE_MODE == PERSISTENCE_RAM) {
+        return g_hash_table_lookup(g_global_block_table, block_header_hash);
     }
-
-    sprintf(sql_query, "select txid from transaction where block_id=%lu;", block_id);
-    res = mysql_read(sql_query);
-    int tx_count = 0;
-    while ((row = mysql_fetch_row(res))) {
-        memcpy(txids[tx_count], row[0], 64);
-        tx_count++;
-    }
-    b->txns = (transaction **)malloc(b->txn_count * sizeof(transaction *));
-    for (int i = 0; i < b->txn_count; i++) {
-        b->txns[i] = get_transaction(txids[i]);
-    }
-
-    mysql_free_result(res);
-    memset(sql_query, 0, temp_sql_query_size);
-    for (int i = 0; i < b->txn_count; i++) free(txids[i]);
-    free(txids);
-
-    return b;
 }
+
+block **get_all_blocks() { return NULL; }
 
 /**
  * Destroy the block persistence layer
@@ -200,8 +264,15 @@ block *get_block(char *block_header_hash) {
  * @author Luke E
  */
 bool destroy_block_persistence() {
-    char *sql_query =
-        "drop table block_header;\n"
-        "drop table block;\n";
-    return mysql_delete_table(sql_query);
+    if (PERSISTENCE_MODE == PERSISTENCE_MYSQL) {
+        char *sql_query =
+            "drop table block_header;\n"
+            "drop table block;\n";
+        return mysql_delete_table(sql_query);
+    } else if (PERSISTENCE_MODE == PERSISTENCE_RAM) {
+        g_hash_table_foreach(g_global_block_table, free_g_global_block_table_entry, NULL);
+        g_hash_table_destroy(g_global_block_table);
+        return true;
+    }
+    return false;
 }
